Delete user

+++ src/main/java/com/example/ticketgui/GUI/Controller/UserController.java
@@ -43,6 +43,8 @@ public class UserController extends Controller {
     private TableColumn<User, UserRole> colRolle;
     @FXML
     private Label lblFeedback;
+    @FXML
+    private Label txtFeedback;
 
 
     @Override
@@ -144,14 +146,15 @@ public class UserController extends Controller {
     private void deleteUser(ActionEvent actionEvent) {
         User selctedUser = tblBrugere.getSelectionModel().getSelectedItem();
         if (selctedUser != null){
-            // TODO : lav et tjek at brugeren ikke sletter sig selv ;)
+            if (ShowAlerts.displayWarning("Deletion of User", "Are you sure that you want to delete this User?:\n" + selctedUser.getName())) {
+                try {
+                    userModel.deleteUser(selctedUser);
+                    txtFeedback.setText("User deleted!");
+                } catch (Exception e) {
+                    txtFeedback.setText("Could not delete the user. Try again later!");
+                }
             // TODO : måske lidt verificering af brugerens valg
-            try{
-                userModel.deleteUser(selctedUser);
-            }
-            catch (Exception e){
-                lblFeedback.setText("Could not delete user");
-                lblFeedback.setStyle("-fx-text-fill: red;");
+
             }
         }
         else{
